reutilize commum helpers
